==> research/01_LLM_ticker.ipynb <==
from phi.agent import Agent
from phi.model.groq import Groq
from phi.model.openai import OpenAIChat
from phi.tools.yfinance import YFinanceTools
from phi.tools.duckduckgo import DuckDuckGo
from datetime import datetime

today = datetime.now().strftime("%B %d, %Y")

web_agent = Agent(
    name="WebAgent",
    role="Market research expert",
    model=Groq(id="meta-llama/llama-4-scout-17b-16e-instruct"),
    tools=[DuckDuckGo()],
    instructions=[
    f"You are a web analyst. Search current news (as of {today}) to identify Indian stocks showing bullish trends today. Focus on companies with positive momentum, news, or sentiment."
    ],
    show_tools_calls=True,
    markdown=True,
    )


finance_agent = Agent(
        name="FinanceAnalyst",
        role="Finance analyst expert",
        model=OpenAIChat(id="gpt-4o"),
        tools=[
            YFinanceTools(
                company_news=True,
                technical_indicators=True,
                historical_prices=True,
                analyst_recommendations=True,
                stock_price=True,
                income_statements=True,
                key_financial_ratios=True,
                company_info=True
            )
        ],
        instructions=[
            f"You are a financial analyst. Study financial data for stocks considered bullish today ({today}) and select the single most promising Indian stock for intraday or short-term trading."
        ],
        show_tools_calls=True,
        markdown=True,
    )

finsight_agent = Agent(
        team=[web_agent, finance_agent],
        tools= [YFinanceTools()],
        instructions=[
            f"Based on the information and research done by your team as of {today}, return only NSE tickers that are available on Yahoo Finance (ending with .NS). Do not suggest BSE or delisted stocks. Do NOT include any other text, explanation, or formatting. Only output the ticker symbol, nothing else."
        ],
        show_tools_calls=True,
        markdown=True,
    )

finsight_agent.run(finsight_agent)


from src.finance_ml.components.LLM_ticker import get_bullish_ticker
from src.finance_ml import logger

class TickerFinderPipeline:
    def main(self):
        logger.info("Running TickerFinderPipeline")
        ticker = get_bullish_ticker()
        logger.info(f"TickerFinderPipeline finished with ticker: {ticker}")
        print("Final Ticker selected:", ticker)
        print("Ticker Type:", type(ticker))

        return ticker
