==> research/03_data_validation.ipynb <==
import os
#os.chdir(r'C:\Projects\FinSight')
%pwd


import pandas as pd

data = pd.read_csv("artifacts/data_ingestion/raw_data.csv")


data.info()

data.describe()

data.isna().sum()

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    STATUS_FILE: str
    data: Path
    all_schemas: dict

from finance_ml.constants import *
from finance_ml.utils.common import read_yaml, create_directories

# src/finance_ml/config/configuration.py

class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        params_filepath=PARAMS_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH
    ):
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_Validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_Validation_config = DataValidationConfig(
            root_dir= config.root_dir,
            STATUS_FILE= config.STATUS_FILE,
            data= config.data,
            all_schemas= schema  
        )

        return data_Validation_config

import os
from finance_ml import logger

class DataValidation:
    def __init__(self, config: DataValidationConfig):
        self.config = config

    def validate_all_columns(self) -> bool:
        try:
            validation_status = True
            data = pd.read_csv(self.config.data)
            all_cols = list(data.columns)
            all_schema = self.config.all_schemas

            for col in all_schema:
                if col not in all_cols: 
                    validation_status = False
                    break

         
            for col, expected_type in all_schema.items():
                if col in data.columns:
                    actual_type = str(data[col].dtype)
                    if actual_type != expected_type:
                        validation_status = False
                        break

            with open(self.config.STATUS_FILE, 'w') as f:
                f.write(f"Validation status: {validation_status}")

            return validation_status
        except Exception as e:
            raise e

try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(config=data_validation_config)
    data_validation.validate_all_columns()
except Exception as e:
    raise e
